test(storage): fold logs/archive checks into the storage-reuse suite
The standalone tests/integration/storage/ suite had no spread task.yaml, so
it never ran in CI. Move the "volumes attached" and "log files present"
checks into tests/integration/ha/test_storage_reuse.py, which already
deploys three units and has k8s+vm spread files, and delete the orphaned
suite.
